Expand on implementation comments
All files without many comments are 'stock' files copied over from the GeoJSON implementation and I honestly don't know what most of them do. Same goes for a lot of the data that comes into and goes out of the functions. Most of it has been cleaned up, but there is probably more to do there.